# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  ds.PaginatedResponse:
    properties:
      items: {}
      total:
        type: integer
    type: object
  ds.ProductCreateRequest:
    properties:
      c_pol:
        type: integer
      n_pol:
        type: integer
      text:
        type: string
      title:
        type: string
    required:
    - text
    - title
    type: object
  ds.ProductDTO:
    properties:
      c_pol:
        type: integer
      id:
        type: integer
      image:
        type: string
      n_pol:
        type: integer
      text:
        type: string
      title:
        type: string
    type: object
...
  ds.ProductUpdateRequest:
    properties:
      c_pol:
        type: integer
      n_pol:
        type: integer
      text:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8090
info:
  contact:
    email: support@example.com
    name: API Support
  description: API-сервер для управления заявками и продуктами в системе DIET.
  title: API для системы DIET
  version: "1.0"
paths:
...
  /diet/draft/products/{product_id}:
    post:
      description: Находит или создает черновик заявки для текущего пользователя и
        добавляет в него продукт.
      parameters:
      - description: ID продукта для добавления
        in: path
        name: product_id
        required: true
        type: integer
      responses:
        "201":
          description: Сообщение об успехе
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Добавить продукт в черновик заявки (все)
      tags:
      - products
...
  /products:
    get:
      description: Возвращает постраничный список продуктов.
      parameters:
      - description: Фильтр по названию продукта
        in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.PaginatedResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список продуктов (все)
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Создает новую запись о продукте.
      parameters:
      - description: Данные нового продукта
        in: body
        name: productData
        required: true
        schema:
          $ref: '#/definitions/ds.ProductCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ds.ProductDTO'
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен (не модератор)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать новый продукт (только модератор)
      tags:
      - products
  /products/{id}:
    delete:
      description: Удаляет продукт риска из системы.
      parameters:
      - description: ID продукта для удаления
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить продукт (только модератор)
      tags:
      - products
    get:
      description: Возвращает детальную информацию о продукте.
      parameters:
      - description: ID продукта
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.ProductDTO'
        "404":
          description: Продукт не найден
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить один продукт по ID (все)
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Обновляет информацию о существующем продукте.
      parameters:
      - description: ID продукта
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/ds.ProductUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.ProductDTO'
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить продукт (только модератор)
      tags:
      - products
  /products/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Загружает и привязывает изображение к продукту.
      parameters:
      - description: ID продукта
        in: path
        name: id
        required: true
        type: integer
      - description: Файл изображения
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: URL загруженного изображения
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Файл не предоставлен
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Загрузить изображение для продукта (только модератор)
      tags:
      - products
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"